Bringing staging into line with develop. Looks like the merge didn't work quite right.
Former-commit-id: 4d95321c31e87dbaaedee2e7b9bfed9115fdb51e

+++ src/main/web/florence/js/functions/_loadTablesList.js
@@ -18,38 +18,44 @@ function loadTablesList(data, collectionId) {
     });
 
     $("#table-delete_" + table.filename).click(function () {
-      $("#table_" + index).remove();
+      var result = confirm("Are you sure you want to delete this table?");
+      if (result === true) {
+        $("#table_" + index).remove();
 
-      getPageData(collectionId, tableJson,
-        onSuccess = function (tableData) {
+        getPageData(collectionId, tableJson,
+          onSuccess = function (tableData) {
 
-          // delete any files associated with the table.
-          _(tableData.files).each(function (file) {
-            var fileToDelete = basePath + '/' + file.filename;
-            deleteContent(collectionId, fileToDelete,
-              onSuccess = function () {
-                //console.log("deleted table file: " + fileToDelete)
-              });
-          });
-
-          // delete the table json file
-          deleteContent(collectionId, tableJson,
-            onSuccess = function () {
-
-              // remove the table from the page json when its deleted
-              data.tables = _(data.tables).filter(function (item) {
-                return item.filename !== table.filename
-              });
-
-              // save the updated page json
-              postContent(collectionId, basePath, JSON.stringify(data),
-                success = function () {
-                  Florence.Editor.isDirty = false;
-                  refreshPreview();
-                  loadTablesList(data, collectionId);
+            // delete any files associated with the table.
+            _(tableData.files).each(function (file) {
+              var fileToDelete = basePath + '/' + file.filename;
+              deleteContent(collectionId, fileToDelete,
+                onSuccess = function () {
+                  //console.log("deleted table file: " + fileToDelete)
                 });
             });
-        });
+
+            // delete the table json file
+            deleteContent(collectionId, tableJson,
+              onSuccess = function () {
+
+                // remove the table from the page json when its deleted
+                data.tables = _(data.tables).filter(function (item) {
+                  return item.filename !== table.filename
+                });
+
+                // save the updated page json
+                postContent(collectionId, basePath, JSON.stringify(data),
+                  success = function () {
+                    Florence.Editor.isDirty = false;
+                    refreshPreview();
+                    loadTablesList(data, collectionId);
+                  }
+                );
+              }
+            );
+          }
+        );
+      }
     });
   });
 }
